// src/components/UI-kit/inputs/InputTextUI/InputTextUI.types.ts
import {InputHTMLAttributes} from 'react'
type TTheme = 'white' | 'dark'
export interface IField extends InputHTMLAttributes<HTMLInputElement> {
  placeholder: string
  error?: string
  extraClass?: string
  type?: string
  extraStyle?: React.CSSProperties
  icon?: React.ReactNode
  labelText?: string
  extraClassClearIcon?: string
  disabled?: boolean
  theme?: TTheme
}